adding metal + brick textures; setting up inventory editors #115
