+++ client/src/js/redux/constants/action-types.js
@@ -153,6 +153,8 @@ export const UTIL_LAUNCH_IMPORT_TO_INVENTORY = 'UTIL_LAUNCH_IMPORT_TO_INVENTORY'
 export const UTIL_LAUNCH_IMPORT_TO_WORLD = 'UTIL_LAUNCH_IMPORT_TO_WORLD'
 export const UTIL_LAUNCH_INVENTORY_EDITOR = 'UTIL_LAUNCH_INVENTORY_EDITOR'
 export const UTIL_LAUNCH_INVENTORY_EXPORT = 'UTIL_LAUNCH_INVENTORY_EXPORT'
+export const UTIL_LAUNCH_COMPONENT_EDITOR = 'UTIL_LAUNCH_COMPONENT_EDITOR'
+export const UTIL_LAUNCH_PROPERTY_EDITOR = 'UTIL_LAUNCH_PROPERTY_EDITOR'
 export const UTIL_CLOSE_TEXT_EDIT = 'UTIL_CLOSE_TEXT_EDIT'
 export const UTIL_CLOSE_RENAME_FILE = 'UTIL_CLOSE_RENAME_FILE'
 export const UTIL_CLOSE_SHARING_SETTINGS = 'UTIL_CLOSE_SHARING_SETTINGS'
@@ -160,6 +162,8 @@ export const UTIL_CLOSE_IMPORT_TO_INVENTORY = 'UTIL_CLOSE_IMPORT_TO_INVENTORY'
 export const UTIL_CLOSE_IMPORT_TO_WORLD = 'UTIL_CLOSE_IMPORT_TO_WORLD'
 export const UTIL_CLOSE_INVENTORY_EDITOR = 'UTIL_CLOSE_INVENTORY_EDITOR'
 export const UTIL_CLOSE_INVENTORY_EXPORT = 'UTIL_CLOSE_INVENTORY_EXPORT'
+export const UTIL_CLOSE_COMPONENT_EDITOR = 'UTIL_CLOSE_COMPONENT_EDITOR'
+export const UTIL_CLOSE_PROPERTY_EDITOR = 'UTIL_CLOSE_PROPERTY_EDITOR'
 
 export const INVENTORY_SET_CURRENT = 'INVENTORY_SET_CURRENT'
 export const INVENTORY_ADD_FETCH = 'INVENTORY_ADD_FETCH'
