+++ client/src/js/util.js
@@ -1,0 +1,7 @@
+export let rgb = ( r, g, b ) => { // because I never remeber to quote that rofl..
+    return `rgb(${r}, ${g}, ${b})`
+}
+
+export let rgba = ( r, g, b, a ) => { 
+    return `rgba(${r}, ${g}, ${b}, ${a})`
+}
